+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -18,6 +18,7 @@ export 'temu.dart';
 export 'tencent_maps.dart';
 export 'threads.dart';
 export 'tiktok.dart';
+export 'tomtom_go_fleet.dart';
 export 'twitter.dart';
 export 'two_gis.dart';
 export 'waze.dart';