+++ test/deeplink_x_test.dart
@@ -191,6 +191,11 @@ void main() {
         expect(action, isA<App>());
       });
 
+      test('TomTom Go Fleet', () {
+        final action = TomTomGoFleet.open();
+        expect(action, isA<App>());
+      });
+
       test('Neshan', () {
         final action = Neshan.open();
         expect(action, isA<App>());